requirements.txt
pandas
python-docx
pytest
